// fronten/app/src/app/models/user.ts
export class User{
    username: string;
    password: string;
    firstname: string;
    lastname: string;
    phone: string;
    email: string;
    orgName: string;
    state : string;
    city : string;
    postNum : number;
    street : string;
    streetNum : number;
    MB: number;
    type: number;
    status: string;
    profileImg : string;
}